+++ src/grandorgue/GOFrame.h
@@ -123,6 +123,7 @@ private:
 
   void OnOrganSettings(wxCommandEvent &event);
   void OnMidiList(wxCommandEvent &event);
+  void OnStops(wxCommandEvent &event);
 
   void OnAudioPanic(wxCommandEvent &event);
   void OnAudioMemset(wxCommandEvent &event);
@@ -158,7 +159,7 @@ private:
   void OnMenuOpen(wxMenuEvent &event);
   void OnCloseWindow(wxCloseEvent &event);
 
-  void OnMidiEvent(const GOMidiEvent &event);
+  void OnMidiEvent(const GOMidiEvent &event) override;
 
   void OnUpdateLoaded(wxUpdateUIEvent &event);
   void OnSetTitle(wxCommandEvent &event);