+++ src/grandorgue/combinations/GOSetter.cpp
@@ -305,7 +305,7 @@ GOSetter::GOSetter(GOOrganController *organController)
     m_CrescendoDisplay(organController),
     m_TransposeDisplay(organController),
     m_NameDisplay(organController),
-    m_swell(*organController) {
+    m_CrescendoCtrl(*organController) {
   CreateButtons(*m_OrganController);
 
   m_buttons[ID_SETTER_PREV]->SetPreconfigIndex(0);
@@ -459,7 +459,7 @@ void GOSetter::Load(GOConfigReader &cfg) {
   m_buttons[ID_SETTER_FILE_EXIT]->Init(
     cfg, wxT("ExitGO"), _("Exit GrandOrgue"));
 
-  m_swell.Init(cfg, wxT("SetterSwell"), _("Crescendo"), 0);
+  m_CrescendoCtrl.Init(cfg, wxT("SetterSwell"), _("Crescendo"), 0);
 
   m_PosDisplay.Init(cfg, wxT("SetterCurrentPosition"), _("sequencer position"));
   m_BankDisplay.Init(cfg, wxT("SetterGeneralBank"), _("general bank"));
@@ -996,7 +996,9 @@ void GOSetter::OnCombinationsSaved(const wxString &yamlFile) {
 
 void GOSetter::Update() {}
 
-void GOSetter::SetterActive(bool on) { m_buttons[ID_SETTER_SET]->Set(on); }
+void GOSetter::SetterActive(bool on) {
+  m_buttons[ID_SETTER_SET]->SetButtonState(on);
+}
 
 void GOSetter::ToggleSetter() { m_buttons[ID_SETTER_SET]->Push(); }
 
@@ -1025,8 +1027,8 @@ void GOSetter::PushGeneral(
     = GetCrescendoAddSet(elementSet);
 
   NotifyCmbPushed(cmb.Push(m_state, pExtraSet));
-  if (!pExtraSet) { // Otherwise the crescendo in add mode:
-                    // not to switch off combination buttons
+  if (pButtonToLight || !pExtraSet) { // Otherwise the crescendo in add mode:
+                                      // not to switch off combination buttons
     UpdateAllSetsButtonsLight(pButtonToLight, -1);
   }
 }
@@ -1042,7 +1044,7 @@ void GOSetter::PushDivisional(
       = GetCrescendoAddSet(elementSet);
 
     NotifyCmbPushed(cmb.Push(m_state, pExtraSet));
-    if (!pExtraSet)
+    if (pButtonToLight || !pExtraSet)
       UpdateAllSetsButtonsLight(pButtonToLight, cmbManual);
   }
 }
@@ -1159,8 +1161,8 @@ void GOSetter::Crescendo(int newpos, bool force) {
 }
 
 void GOSetter::ControlChanged(GOControl *control) {
-  if (control == &m_swell)
-    Crescendo(m_swell.GetValue() * CRESCENDO_STEPS / 128);
+  if (control == &m_CrescendoCtrl)
+    Crescendo(m_CrescendoCtrl.GetValue() * CRESCENDO_STEPS / 128);
 }
 
 void GOSetter::UpdateTranspose() {
@@ -1186,7 +1188,7 @@ void GOSetter::UpdateModified(bool modified) {
 
 GOEnclosure *GOSetter::GetEnclosure(const wxString &name, bool is_panel) {
   if (name == wxT("Swell"))
-    return &m_swell;
+    return &m_CrescendoCtrl;
 
   return NULL;
 }