+++ src/grandorgue/model/GOCoupler.cpp
@@ -15,8 +15,10 @@
 #include "GOManual.h"
 #include "GOOrganModel.h"
 
-GOCoupler::GOCoupler(GOOrganModel &organModel, unsigned sourceManual)
+GOCoupler::GOCoupler(
+  GOOrganModel &organModel, unsigned sourceManual, bool isVirtual)
   : GODrawstop(organModel),
+    m_IsVirtual(isVirtual),
     m_UnisonOff(false),
     m_CoupleToSubsequentUnisonIntermanualCouplers(false),
     m_CoupleToSubsequentUpwardIntermanualCouplers(false),
@@ -357,7 +359,7 @@ void GOCoupler::SetKey(
   ChangeKey(note, velocity);
 }
 
-void GOCoupler::ChangeState(bool on) {
+void GOCoupler::OnDrawstopStateChanged(bool on) {
   if (m_UnisonOff)
     r_OrganModel.GetManual(m_SourceManual)->SetUnisonOff(on);
   else {